+++ docs/conf.py
@@ -14,7 +14,7 @@ import cv2ext
 project = 'cv2ext'
 copyright = '2024, Justin Davis'
 author = 'Justin Davis'
-version = '0.0.13'
+version = '0.0.14'
 
 assert version == cv2ext.__version__