+++ src/cv2ext/cli/video_from_images/__main__.py
@@ -1,0 +1,11 @@
+# Copyright (c) 2024 Justin Davis (davisjustin302@gmail.com)
+#
+# MIT License
+"""Create a video from a directory of images."""
+
+from __future__ import annotations
+
+from ._convert import video_from_images_cli
+
+if __name__ == "__main__":
+    video_from_images_cli()